requirements.txt
numpy
jax
dm-haiku
optax
